Allow to specify a seed to use for random operators (#29).
Seed can be specified when launching an experiment. If none, a random
one will be generated. This seed will be used to generate a new seed
for each run, that will then be made available to operators.

Inside operators, the actual behavior is to generate one last seed
per trace, thus allowing a deterministic behavior even in a multi-core
settings.

Tests have been updated to use a fixed seed that "just works".

GitOrigin-RevId: 0b1e6e17bc716d1e5d10e59e92b6899b4244cc08