Split the Accio project between Accio and Lumos
Accio is now only in charge of the workflow execution, while Lumos
is basically an experiments database. Lumos provides a data model
generic enough to be used in other contexts. Lumos follows an
event-based model, where events are pushed and the state is updated
accordingly.

The execution model of Accio has also been modified: the executor is
now in charge of handling an entire workflow (instead of a single
operator as it was the case until now). We lose a bit of flexibility,
but it matches more closely our use cases, where resources are
determined on a per-workflow basis rather than on a per-operator basis.

Consequently, there are now two CLI tools. The Lumos CLI is read-only,
while the Accio CLI allows to submit workflows.

Lots of tests have been added. Property-based tests have been introduced
where relevant.