Change the way to handle unresolved data types
Previously they were considered as their parent type if the correct type
was not available, which means some information was lost in the process.
Now there is a Value.Unresolved class which is essentially a wrapper
around a value of the parent type, but keeps a reference to the actual
(unresolved) data type name.