Fix an issue with invalid paths
URIs are usually formatted like scheme://authority/path. However, if the
authority is absent, the // part is usually removed, leaving URIs such
as scheme:/path. This may be confusing, as the JDK favours the first
form when generate URIs, but does not seem to accept it everywhere.
This commit introduces some hacks to handle this properly.